#65 - Implementar a opção de desconto nos produtos #66 - Implementar a opção de acréscimo nos produtos #67 - Implementar a opção de alterar preço dos materiais

+++ app/src/main/java/br/com/informsistemas/forcadevenda/controller/adapter/MovimentoParcelaAdapter.java
@@ -48,7 +48,7 @@ public class MovimentoParcelaAdapter extends RecyclerView.Adapter<MovimentoParce
         FormaPagamento p = FormaPagamentoDAO.getInstance(context).findByIdAuxiliar("codigoforma", fList.get(position).codigoforma);
 
         myViewHolder.txtDescricao.setText(p.descricao);
-        myViewHolder.txtValor.setText(Misc.formatMoeda(fList.get(position).valor));
+        myViewHolder.txtValor.setText(Misc.formatMoeda(fList.get(position).valor.floatValue()));
 
         for (int i : selectedIds){
             if (fList.get(position).id == i){
